requirements.txt
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
h5py>=3.9.0
opencv-python>=4.8.0
scipy>=1.11.0
reportlab>=4.0.0
Pillow>=10.0.0
PyYAML>=6.0
pytest>=7.4.0
streamlit>=1.30.0
PyQt5>=5.15.0
matplotlib>=3.8.0
